search function fully fixed
